Kind of working dual OPL2 playback using EMU8950_LINEAR/SLOT_RENDER

+++ sw/opl/emu8950.h
@@ -189,7 +189,7 @@ int16_t OPL_calc(OPL *opl);
 
 void OPL_calc_buffer(OPL *opl, int16_t *buffer, uint32_t nsamples);
 // LE left/right channels int16:int16
-void OPL_calc_buffer_stereo(OPL *opl, int32_t *buffer, uint32_t nsamples);
+void OPL_calc_buffer_stereo(OPL *opl, int16_t *buffer, uint32_t nsamples);
 
 /**
  *  Set channel mask
